JSONCompactEachRow and JSONCompactEachRowWithNamesAndTypes united
